refactor: rename utils into user
